perf: use mmap, read records faster!

+++ src-tauri/src/engine/data/cursor_position.rs
@@ -1,5 +1,5 @@
 pub enum CursorPosition {
     Start,
     End,
-    Middle(u64),
+    Middle(usize),
 }
